+++ CHANGES.rst
@@ -9,6 +9,7 @@ Unreleased Changes
   version in under ``biweeklybudget.vendored.wishlist``.
 * `Issue #113 <https://github.com/jantman/biweeklybudget/issues/113>`_ - vendor in other
   git requirements (ofxclient and ofxparse) that seem unmaintained or inactive, so we can install via ``pip``.
+* `Issue #115 <https://github.com/jantman/biweeklybudget/issues/115>`_ - In Transactions view, add ability to filter by budget.
 
 0.4.0 (2017-08-22)
 ------------------